requirements.txt
cassandra-driver==3.29.1
psycopg2==2.9.9
pymongo~=4.6.3
py2neo~=2021.2.4
nltk~=3.8.1
pandas~=2.2.2
